pyproject.toml
[project]
name = "glacier_mapping"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "numpy<2",
  "addict",
  "geopandas",
  "numba",
  "pandas",
  "pyyaml",
  "rasterio",
  "scikit-image",
  "scikit-learn",
  "seaborn",
  "tensorboard",
  "torch",
  "torchvision",
  "tqdm",
  "elasticdeform",
  "earthaccess",
  "pyproj",
  "netcdf4",
  "xarray",
  "rioxarray",
  "earthengine-api",
  "zarr",
  "s3fs",
  "segmentation-models-pytorch",
  "opencv-python",
  "matplotlib>=3.10.7",
  "pytorch-lightning>=2.0.0",
  "mlflow>=2.0.0",
  "hydra-core>=1.3.0",
  "torchmetrics>=1.0.0",
]

[project.optional-dependencies]
dev = ["ruff", "ipykernel"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["glacier_mapping"]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "ruff>=0.14.6",
]
